Changed Go_SoundD0 to Go_SpecSoundIndex
This is a part of making the special SFXes look less like a cheap hack
made just for $D0, and more of its own system, as it actually is. It
just happens to be occupied by only one sound, and its index just
happens to be underneath the normal SFXes index.