Add build tools and scripts for Linux (x86-64)
A binary of AS 1.42 build 92 for x86-64 is provided (the latest build at
the time of this commit, build 97, gives strange errors), along with the
required message files. Also, a binary of s1p2bin for x86-64 is
provided, along with a makefile to rebuild it. Finally, the Python build
scripts have been updated to use the provided binaries and to calculate
the checksum directly in Python instead of using the fixheader tool
(which remains in the repository because build.bat still uses it); this
should work on all platforms. The Python build scripts require Python 3.