Make out_of_range size explicit
Without an explicit size, the macro was being expanded to include the
line 'bhi. exit', which just looks weird, and doesn't make it clear
to the reader what the size will actually be when assembled. asm68k
assumes word-size in this case, but the reader shouldn't be expected
to know that.