Restore the pre-calculated checksums
This was removed in 3aa7891a2bee2c3bfc1a38e21257bb273f99bc2c, but I'm
restoring it because, this way, the accuracy of the built ROM can be
verified without needing to run fixheader.exe first.

This is useful for situations where fixheader.exe can't be ran (or is
just annoying to run), such as when using a non-Windows system.

Yeah, yeah, I know, asm68k is Windows-only. That doesn't mean that
the disassembly can't be assembled with something else though.

Cough, cough.