Replace the Batch and Python scripts with Lua.
This was done to reduce duplication, as the Batch and Python build
scripts both do the same thing albeit in completely different ways,
making maintaining the build system a pain.

The two build scripts are for different platforms: Batch is for
Windows users while Python is for Linux and Mac users.

Since Python is cross-platform, you might be wondering why we can't
just make Windows users use the Python script. The reason is that
Windows users like being able to build the disassembly without
installing anything. We could bundle an entire Python runtime with
the disassembly so that it doesn't need installing separately, but
Python's runtime is massive and would hugely bloat the repo.

Lua, on the other hand, has a tiny runtime: the version in this
commit is only 500KB. This makes it more than light enough to bundle
with the disassembly for Windows users.

With a runtime bundled, it becomes feasible for Windows users to use
a Lua-based build script without installing anything. But, of course,
Windows users also hate using command line tools: they just want to
double-click on a file to build the disassembly. To address this, the
Batch build script has been kept, but hollowed-out and reduced to a
mere shim to run the Lua script. To the user, the process of building
the ROM won't have changed at all, and it is as intuitive as ever.

Likewise, to Linux and Mac users, the process should also be as
simple as it was before: like with the the Python script, the Lua
script expects Linux and Mac users to install a Lua runtime, rather
than depend on one being bundled with the disassembly.

The Lua script has been given an appropriate shebang, so that it can
be directly executed on Linux (and Mac?) much like how the Batch
script can be directly executed on Windows.

Because Lua is a powerful language, I've merged the functionality of
the `fixheader.exe` tool into the Lua build script, just like the
Python build script. This means that the Lua script computes the
checksum and patches the ROM header on its own, without using a
native executable as the Batch script did.

A Lua-based 'chkbitperfect' script has been provided as well,
replacing the previous two Batch scripts and the similar
functionality in the Python build script.

Fun fact: I couldn't find any fast Lua-based SHA256 or MD5 hashers,
so I converted my own C89 MD5 hasher to Lua, and it worked a treat!

My plan is to migrate the Sonic 2 and Sonic 3 & Knuckles
disassemblies to use a Lua build system as well. This would eliminate
the fragmentation caused by each disassembly having two build
systems, allowing for the build process of each disassembly to be
made more elaborate without the burden of maintaining two build
systems.

For instance, I'd like to get the Sonic 2 disassembly to use ASM
music, but to do this I'd need to add a step to the build process
where each song is assembled and compressed before the ROM itself is
built. While this idea is daunting when considering the fact that I'd
have to implement it *twice* - once for the Batch build system and
again for the Bash build system - it's not a problem at all when it's
just one build system. It helps that I'm a lot more familiar with Lua
than I am with Batch and Bash.

Oh, by the way, the build script now defaults to the accurate DAC
driver compression: users that want improved compression can just
edit the boolean at the start of the script.